make loop variable size_t as well when looping to a size_t limit

+++ src/getpass.c
@@ -21,7 +21,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * ============================================================================
  *
- * $Id: getpass.c,v 1.5 2004-02-02 07:24:30 bagder Exp $
+ * $Id: getpass.c,v 1.6 2004-03-10 16:03:12 bagder Exp $
  *
  * The spirit of this license is to allow use of this source code in any
  * project be it open or closed but still encourage the use of the open,
@@ -199,7 +199,7 @@ char *getpass_r(const char *prompt, char *buffer, size_t buflen)
 #include <conio.h>
 char *getpass_r(const char *prompt, char *buffer, size_t buflen)
 {
-  int i;
+  size_t i;
   printf("%s", prompt);
  
   for(i=0; i<buflen; i++) {
